fix(native): own backgroundBlur scratch buffers on the toolkit instead of leaking per-thread
